+++ R/workspace.R
@@ -206,6 +206,7 @@ get_context <- function(jws) {
 #'
 #' Functions to read all the SAItem of a SAProcessing (`read_sap()`)
 #' or a workspace (`read_workspace()`).
+#' The functions `.jread_sap()` and `.jread_workspace()` only returns the Java objects
 #'
 #' @param jws Java workspace.
 #' @param jsap Java SAProcessing.
@@ -229,6 +230,20 @@ read_workspace <- function(jws, compute = TRUE) {
     names(jsaps) <- names
     cntxt <- get_context(jws)
     return(list(processing = jsaps, context = cntxt))
+}
+#' @name read_workspace
+#' @export
+.jread_workspace <- function(jws, compute = TRUE) {
+    if (compute) .jws_compute(jws)
+    n <- .jws_sap_count(jws)
+    jsaps <- lapply(1:n, function(i) {
+        .jread_sap(.jws_sap(jws, i))
+    })
+    names <- lapply(1:n, function(i) {
+        .jsap_name(.jws_sap(jws, i))
+    })
+    names(jsaps) <- names
+    return(jsaps)
 }
 
 #' Save Workspace