+++ NEWS.md
@@ -9,6 +9,8 @@ to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 * fixed `set_specification()` and `set_domain_specification()`
 
+* new function `.jsa_jresults()`, `.jread_sap()` and `.jread_workspace()` to  extract the Java object of the results of a SAItem.
+
 ## [3.2.3] - 2024-07-12